Separating declarations to separate file
